feat: dataset title permission

+++ apis/dataset_metadata/dataset_title.py
@@ -42,9 +42,19 @@ class DatasetTitleResource(Resource):
         for i in data:
             if "id" in i and i["id"]:
                 dataset_title_ = model.DatasetTitle.query.get(i["id"])
+                if dataset_title_.type == "MainTitle":
+                    return (
+                        "MainTitle type can not be modified",
+                        403,
+                    )
                 dataset_title_.update(i)
                 list_of_elements.append(dataset_title_.to_dict())
             elif "id" not in i or not i["id"]:
+                if i["type"] == "MainTitle":
+                    return (
+                        "MainTitle type can not be given",
+                        403,
+                    )
                 dataset_title_ = model.DatasetTitle.from_data(data_obj, i)
                 model.db.session.add(dataset_title_)
                 list_of_elements.append(dataset_title_.to_dict())
@@ -69,6 +79,11 @@ class DatasetTitleResource(Resource):
                     403,
                 )
             dataset_title_ = model.DatasetTitle.query.get(title_id)
+            if dataset_title_.type == "MainTitle":
+                return (
+                    "MainTitle type can not be deleted",
+                    403,
+                )
             model.db.session.delete(dataset_title_)
             model.db.session.commit()
             return 204
